Structured the frontend code into modules
