+++ static/downloadingWebsite/js/utils/screen.js
@@ -1,0 +1,7 @@
+export const screen = {
+    getDeviceScreenWidth()
+    {
+        let width = (window.innerWidth > 0) ? window.innerWidth : screen.width;
+        return width;
+    }
+}
